+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.5.1 (2025-01-22)
+
 - changed: Remove Blockbook servers from persisted internal servers list when updating this list from payload/default servers.
 - fixed: Remove blockbook.firo.org Blockbook server form default Blockbook server list.